% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIDIFAC.R
\name{BIDIFAC}
\alias{BIDIFAC}
\title{BIDIFAC}
\usage{
BIDIFAC(
  data,
  rmt = T,
  sigma = NULL,
  start = NULL,
  out = FALSE,
  eps = 0.001,
  max.iter = 1000,
  pbar = TRUE,
  seed = NULL,
  scale_back = TRUE,
  ...
)
}
\description{
BIDIFAC
}
\author{
Jun Young Park (2020)
}
