% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp.R
\name{bsfp}
\alias{bsfp}
\title{Bayesian Simultaneous Factorization and Prediction (BSFP)}
\usage{
bsfp(
  data,
  Y,
  nninit = TRUE,
  model_params = NULL,
  ranks = NULL,
  scores = NULL,
  nsample,
  burnin = NULL,
  progress = TRUE,
  starting_values = NULL
)
}
\arguments{
\item{data}{A matrix of lists or a list of matrices that share the same number of
columns. The matrices must be oriented in pxn orientation. May contain NAs if
there are missing values in the dataset.}

\item{Y}{A matrix of lists or a nx1 matrix of continuous or binary outcome.
May be NULL if no outcome is given. May contain NAs if there are missing outcomes.}

\item{nninit}{Boolean determining if nuclear-norm penalized objective is used
to initialize the model. If TRUE, ranks = NULL.}

\item{model_params}{A list of hyperparameters for the model.
May be left NULL if theoretical defaults are desired. Otherwise, must be in
the following named-list form: model_params = list(error_vars = c(), joint_var = double(),
indiv_vars = c(), beta_vars = c(), response_vars = c()). error_vars, indiv_vars are
vectors of length q for each source. response_vars must define the shape and rate for
the Inverse-Gamma prior of the response variance. beta_vars must be of length q+1
to specify the prior variance on the intercept, the prior variance on each joint factor's
contribution to the outcome, and for each individual factor's contribution from each source.}

\item{ranks}{A list of length q+1 for the ranks of the joint and individual structures.
Leave NULL if nninit=TRUE.}

\item{scores}{Matrix with scores estimated by existing factorization method. Use only
if desired to run Bayesian linear model with scores estimated separately. Otherwise,
leave NULL.}

\item{nsample}{Integer specifying the number of posterior samples to generate.}

\item{burnin}{Integer specifying the number of posterior sampling iterations to burn.
Default is nsample/2, which is used if burnin=NULL.}

\item{progress}{Should a progress bar be displayed to visualize the progress of the sampler?}

\item{starting_values}{List of initial values for Gibbs sampler. If NULL and nninit=TRUE,
fixes at posterior mode. If NULL and nninit=FALSE, simulates from prior distributions.}
}
\description{
Given q sources of data and a continuous or binary outcome measured on n samples,
BSFP can decompose variation across the sources and use the estimate factors to
predict the outcome.
Use this function to simulate samples from posterior distributions
of joint and individual structures. This function
can be used in several ways:
(1) Initialize at the theoretical mode of the decomposition. Priors are fixed
at theoretical values. Ranks are determined at initialization. Can include an
outcome here or not. If no outcome is included, no burn-in is used. If an outcome
is included, a burn-in must be specified or the default is nsample/2.
(2) User specifies ranks to estimate. Model is initialized using priors and
then function samples from posterior distributions. User must specifiy the
hyperparameters for the prior distributions on the structures.
}
