% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp.R
\name{bsfp_data}
\alias{bsfp_data}
\title{bsfp_data}
\usage{
bsfp_data(
  p.vec,
  n,
  ranks,
  true_params,
  s2nX = NULL,
  s2nY = NULL,
  response = NULL,
  missingness = NULL,
  missing_data_type = NULL,
  prop_missing = NULL,
  sparsity,
  identically_zero = FALSE,
  num_in_spike = NULL
)
}
\arguments{
\item{p.vec}{(vector) vector of with number of features per source}

\item{n}{(int) number of samples}

\item{ranks}{(vector) vector with joint and individual ranks (first index
is joint, remaining are individual for each source)}

\item{s2nX}{(dbl) signal-to-noise ratio in X (ratio of variance in X to variance
in error). >1 means higher signal in X, <1 means lower signal in X}

\item{s2nY}{(dbl) signal-to-noise ratio in Y}

\item{response}{NULL if no response is desired, "continuous", or "binary"}

\item{missingness}{NULL if no missingness, or "missingness_in_data" or "missingness_in_response"}

\item{missing_data_type}{NULL if missingness = NULL; otherwise, = entrywise
if randomly sample across sources; columnwise if randomly set samples to missing;
 MNAR is randomly set features below a threshold to missing}

\item{prop_missing}{NULL if missingness = NULL; otherwise the proportion of
entries or columns set to missing}

\item{sparsity}{(Boolean) TRUE if generate regression coefficients under
spike-and-slab prior, FALSE otherwise}

\item{identically_zero}{(Boolean) TRUE if generating response with sparsity
and want spike coefficients to be exactly 0}

\item{num_in_spike}{(vector) number of coefficients to be generated from spike
if generating response with sparsity. NULL if sparsity=FALSE. Should have an
integer for joint and each individual structure. May be set to NULL and
the number of factors in the spike will be determined randomly.}
}
\description{
Generate fake data according to BSF/BSFP model
}
\details{
Generate data according to the BSF and BSFP models. Use for examples or
simulations to assess model performance.
}
