% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIDIFAC.R
\name{impute.BIDIFAC}
\alias{impute.BIDIFAC}
\title{impute.BIDIFAC}
\usage{
impute.BIDIFAC(
  data,
  rmt = T,
  sigma = NULL,
  pbar = TRUE,
  start = NULL,
  max.iter.impute = 100,
  eps.impute = 0.001,
  scale_back = TRUE,
  ...
)
}
\description{
impute.BIDIFAC
}
\author{
Jun Young Park (2020)
}
