% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp.R
\name{log_joint_density}
\alias{log_joint_density}
\title{log_joint_density:}
\usage{
log_joint_density(
  data,
  Y = NULL,
  U.iter,
  V.iter,
  W.iter,
  Vs.iter,
  model_params,
  ranks,
  beta.iter = NULL,
  tau2.iter = NULL,
  Xm.iter = NULL,
  Ym.iter = NULL
)
}
\arguments{
\item{data}{A matrix of lists or a list of matrices that share the same number of
columns. The matrices must be oriented in pxn orientation. May contain NAs if
there are missing values in the dataset.}

\item{Y}{A matrix of lists or a nx1 matrix of continuous or binary outcome.
May be NULL if no outcome is given. May contain NAs if there are missing outcomes.}

\item{U.iter}{Posterior draw for joint loadings, U, at a given iteration}

\item{V.iter}{Posterior draw for joint scores, V, at a given iteration}

\item{W.iter}{Posterior draw for individual loadings, W, at a given iteration}

\item{Vs.iter}{Posterior draw for individual scores, Vs, at a given iteration}

\item{ranks}{Estimated joint and individual ranks from BSFP}

\item{beta.iter}{Posterior draw for regression coefficients at a given iteration}

\item{tau2.iter}{Posterior draw for estimated error variance in Y at a given iteration}

\item{Xm.iter}{Imputed values for unobserved X values at a given iteration}

\item{Ym.iter}{Imputed values for unobserved Y values at a given iteration}
}
\description{
Calculate the log-joint density of the estimated model at a given posterior
sampling iteration to facilitate checking convergence.
}
