% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_align_bsfp.R
\name{match_align_bsfp}
\alias{match_align_bsfp}
\title{match_align_bsfp}
\usage{
match_align_bsfp(
  BSFP.fit,
  y = NULL,
  model_params,
  p.vec,
  iters_burnin,
  piv.list = NULL,
  index = 0
)
}
\arguments{
\item{BSFP.fit}{(list): a list of results from the BPMF model fit}

\item{y}{(matrix of lists): outcome vector if available}

\item{model_params}{(list): model parameters used in model fitting}

\item{p.vec}{(vector): vector with number of variables in each source}

\item{iters_burnin}{(vector): indices for posterior samples after burn-in}

\item{piv.list}{(list of matrices or ints): list of pivot matrices for joint then
individual structures or indices for pivots from posterior samples}

\item{index}{(int): for sensitivity of the results, choose a pivot that falls
the some indices away from the chosen pivot (index specifies the number of indices)}
}
\description{
Correcting for rotation, permutation, and sign invariance by applying the
MatchAlign algorithm (Poworoznek et al. (2021)) to results from BSFP fit.
}
