% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp.R
\name{var_explained}
\alias{var_explained}
\title{var_explained}
\usage{
var_explained(BSFP.fit, iters_burnin = NULL, source.names)
}
\arguments{
\item{BSFP.fit}{Output from the BSFP function}

\item{iters_burnin}{(vector) If adding a burn-in, add the iterations after burn-in. May also include thinning.
Default is NULL if burn-in is already added.}

\item{source.names}{(vector) Vector of source names. Should be of length \eqn{q}, the number of sources.}
}
\value{
Returns a list of posterior summaries for the variance explained by the joint and individual structures in each source.
\item{Joint}{Named list with each entry corresponding to the posterior mean and 95\% credible interval for the proportion
of variance explained by the joint structure in each data source. }
\item{Individual}{Named list with each entry corresponding to the posterior mean and 95\% credible interval for the proportion
of variance explained by the individual structures in the corresponding data source. }
\item{Joint.Response}{Vector with the posterior mean and 95\% credible interval for the proportion of variance
explained by the joint factors in the predictive model.}
\item{Individual.Response}{Named list with each entry corresponding to the posterior mean and 95\% credible interval for the proportion of variance
explained by the individual factors in the predictive model.}
}
\description{
Calculates the proportion of variance explained in each data source by each
estimated structure.
}
\details{
Calculate the proportion of variation in each data source by the estimated structures. Provides posterior
summaries of the variance explained.
}
\examples{
# Setting up the data
n <- 50
p.vec <- c(75, 100)
q <- 2

# Setting up the model parameters
true_params <- list(error_vars = c(1,1),
                    joint_var = 1,
                    indiv_vars = c(1,1),
                    beta_vars = c(1, 1, rep(1, q)),
                    response_vars = c(shape = 1, rate = 1))

# Choose ranks
r <- 3
r.vec <- c(3, 3)
ranks <- c(r, r.vec)

# Number of posterior sampling iterations
nsample <- 1000
burnin <- nsample/2
iters_burnin <- (burnin+1):nsample

# Generate data
data.c1 <- bsfp_data(p.vec, n, ranks, true_params, s2nX = NULL, s2nY = NULL, response = "continuous", sparsity = FALSE)

# Run BSFP for 1000 iterations
bsfp.c1 <- bsfp(data = data.c1$data, Y = data.c1$Y, nsample = nsample)

bsfp.c1.var.explained <- var_explained(BSFP.fit = bsfp.c1, iters_burnin = iters_burnin, source.names = c("Expression", "Methylation"))
}
