% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfp.R
\name{bsfp.predict}
\alias{bsfp.predict}
\title{bsfp.Predict}
\usage{
bsfp.predict(
  bsfp.fit,
  test_data,
  Y_test,
  nninit = TRUE,
  model_params = NULL,
  sparsity = FALSE,
  nsample,
  progress = TRUE,
  starting_values = NULL
)
}
\arguments{
\item{Y_test}{Column vector with outcome on test samples or NULL}

\item{nninit}{Should the model be initialized with a nuclear norm penalized
objective? If FALSE, provide ranks.}

\item{model_params}{(error_vars, joint_vars, indiv_vars, beta_vars = NULL, response_vars).
May be given by BSFP.fit}

\item{nsample}{Number of Gibbs sampling iterations}

\item{progress}{Should the progress of the sampler be displayed?}

\item{starting_values}{list of starting values for V, U, W, Vs. If NULL and
nninit = TRUE, init with UNIFAC. If NULL and nninit = FALSE, init from prior.
If not NULL, will init with provided starting values unless nninit.}

\item{bpmf.fit}{Results from a training bpmf.fit}
}
\description{
Gibbs sampling algorithm for sampling the underlying structure and the
regression coefficient vector for a response vector for unseen test data.
}
